requirements.txt
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.3
joblib==1.4.2
kaleido==0.2.1
kiwisolver==1.4.7
llvmlite==0.43.0
matplotlib==3.10.0
numba==0.60.0
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
pynndescent==0.5.13
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
six==1.17.0
tenacity==9.0.0
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2024.2
umap-learn==0.5.7
